[BOJ] 연결 요소의 개수 / S2 / 10분 / DFS 버전
https://www.acmicpc.net/problem/11724
